backend/requirements.txt
asgiref==3.8.1
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
charset-normalizer==3.4.0
# cloudinary==1.41.0
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
Django==5.1.3
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
# djongo==1.3.6
dnspython==2.7.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
mux_python==3.19.0
# publitio==1.3.1
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pymongo==4.10.1
pyOpenSSL==24.2.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
# requests==2.32.3
# sendgrid==6.11.0
service-identity==24.2.0
six==1.16.0

#sqlparse==0.5.1
sqlparse==0.3.1

starkbank-ecdsa==2.2.0
# stripe==11.2.0
tomli==2.1.0
Twisted==24.10.0
txaio==23.1.1
typing_extensions==4.12.2
urllib3==2.2.3
websocat==1.13.0
# websockets==14.1
zope.interface==7.1.1

pusher==3.3.2


google-cloud-storage==3.1.0

modal==0.73.85

gunicorn==23.0.0